implement a sub command mechanism for uiautomator
Currently uiautomator is a simple test runner. To consolidate
multiple functions into it, the command is refactored with a sub
command mechanism that allows more extensions.

Existing UiAutomator class is changed into a sub command
implementation RunTestCommand. A new Launcher class is added
as a generic launcher for various subcommands and help info. The
uiautomator shell script also have some changes to special-case
the 'runtest' subcommand, mainly for things that cannot be
achieved in Java code (setting signal trap and classpath)

Change-Id: Ib3e3b4a5c101fedf396e8094d5734803affa03fa