move UI Automator Viewer, DummyIME
UI Automator Viewer is a tool used to retrieve matching UI
hierarchy dump and XML from device. In addition to moving
the source code, it's also modified to call the new
'uiautomator' command in system partition.

DummyIME is a testing fixture. It's an empty implementation
of an input method. This is used for preventing soft keyboard
from obscuring app window, to make life eaiser for some test
cases.

Change-Id: I13ee07180f5c685fef223ee9ba13bad3150ab03d