add API to get screen size in dp, and a skeleton test
As a replacement for UiDevice#isPhone, a better way to determine
the layout in use is to look at screen size, as most apps are
doing to adjust layout, more specifically, the width in dp

The skeleton test is meant to be used as a prototype for
spawning new UI tests. It has basic setup to be built in the
source tree, and has one single test case.

Change-Id: I1c4738eba4aca24fcb960dfd873c51107777789e