Clipped widget bounds reported outside screen bounds
This change modifies what the tests see in terms of test object bounds
and it also fixes what is reported in the viewer.

Objects used to report whatever widget bounds accessibility reported
and those sometimes streched off the current display size causing some
clicks to be calculated off screen and thus ignored. Clipping to the
display region will result in coordinates that tests can works with.

Change-Id: I56530315ddc7241eb6b0ef7f0d5cc532c9259a78