unhide getDisplaySizeDp, also make it return real size
getDisplaySizeDp used to return the display size available to
apps, which could be the real size minus any system decorations
like status bar. This caused some metrics to change when device
is in different orientations. Since the purpose of the API is
to determine the screen size class, we now return metrics based
on real display sizes.

Change-Id: I820f8321df941650e401adc1435e0da5a27addf8