UiObject did not perform the wait waitForObject() in UiWatchers
The UiWatcher in test libraries use the UiObject#waitForObject
to wait for certain conditions to come true, however polling
for object state was unecessarly disallowed. Initially the
rational was to prevent the UiWatcher code from blocking too
long, but this doesn't prevent the UiWatcher from using
system sleep and blocking anyway.

Change-Id: Icd086e696f903ddc409ce8eabaa87c99648a90b5