Workaround for (provided scope broken for local jars) 2
Added stripping of guava and dagger from classes.jar

Previously only /libs was removed but .class files
for dagger and guava remained in the .aar, which was
not an issue during the build/test but bloated the apk.
The jarjar'd versions of dagger and guava are of course
still there.

Change-Id: I70eee2e6aee355e1342df29b60e82a32a40b2399