Add jarjar rules to shade espresso deps.
This applies the same rules as in the gradle build, shading dagger
and guava dependencies.

Without this, trying to test an app that bundles a different
version of guava to espresso causes odd failures as the same
classes are loaded from different version by the app and the
instrumentation test.

Change-Id: I4fa65f8e12c7924bdae4bd0dbb848b32c136e335